FPGA: Max sample rate was fixed.
This patch fixes the bug in the strobe generator for vita RX chain. If CIC
decimation rate equal to 0 or 1 and halfband filters are disabled, then strob generator also disabled.

This is why UmTRX could work with max sample rate less then half of LMS clock (13Msps/2=6.5Msps).

The solution is that when CIC decimation rate equal to 0 or 1 and halfband filters are disabled strob generator must forming strobe impulse equal to LMS clock.